Spamming signals to comsumers to avoid deadlock waiting
Testing different number of threads of consumers and producers, having an infinite loop in the consumer side, requires condition termination. Due to multithreading, we aren't able to properly detect and inform consumers that there aren't producers so don't wait. To solve the issue we spamming condition signals to the consumers until they terminate